# contracts/openapi/alignment-v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Traffic Analysis Platform Alignment API",
    "version": "1.2.0",
    "description": "Additive contract for the four remediation vertical slices. Existing APIs remain available during migration."
  },
  "servers": [{"url": "/api"}],
  "paths": {
...
    "/internal/v1/audit/batches": {
      "post": {
        "operationId": "ingestAuditLogBatch",
        "x-required-scope": "audit:write",
        "x-feature-id": "F-AUDIT-001",
        "x-rollout-flag": "AUDIT_BATCH_FAIL_CLOSED_V1_ENABLED",
        "x-compatibility": "Additive internal ingress; existing /audit-log UI and audit command routes remain available",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/AuditLogBatchIngest"}
            }
          }
        },
        "responses": {
          "202": {
            "description": "Kafka acknowledged the complete immutable batch; PostgreSQL and ClickHouse projection remains non-final and queryable by event status URL",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Envelope"}}}
          },
          "400": {"description": "The entire batch is rejected when any event or command field is invalid"},
          "401": {"description": "Authenticated tenant identity required"},
          "403": {"description": "audit:write scope required"},
          "503": {"description": "Ingress disabled or Kafka did not acknowledge; no 202 acceptance is returned"}
        }
      }
    },
...
  },
  "components": {
...
    "schemas": {
      "AuditLogBatchIngest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["action_id", "reason", "events"],
        "properties": {
          "action_id": {"const": "audit-batch-ingest"},
          "reason": {"type": "string", "minLength": 3, "maxLength": 1000},
          "events": {
            "type": "array",
            "minItems": 1,
            "maxItems": 200,
            "items": {"$ref": "#/components/schemas/AuditLogIngestEvent"}
          }
        }
      },
      "AuditLogIngestEvent": {
        "type": "object",
        "additionalProperties": false,
        "required": ["event_id", "action", "created_at"],
        "properties": {
          "event_id": {"type": "string", "minLength": 1, "maxLength": 200},
          "tenant_id": {"type": "string", "description": "Optional assertion that must match the authenticated tenant; it never overrides authenticated identity"},
          "user_id": {"type": "string", "maxLength": 200},
          "action": {"type": "string", "minLength": 1, "maxLength": 200},
          "object_type": {"type": "string", "maxLength": 200},
          "object_id": {"type": "string", "maxLength": 512},
          "detail": {},
          "ip_addr": {"type": "string", "maxLength": 128},
          "user_agent": {"type": "string", "maxLength": 2000},
          "created_at": {"type": "integer", "format": "int64", "minimum": 1, "description": "Stable Unix millisecond timestamp included in idempotency identity"}
        }
      },
...
      "Envelope": {
        "type": "object",
        "required": ["data", "meta", "error"],
        "properties": {
          "data": {},
          "meta": {"$ref": "#/components/schemas/Meta"},
          "error": {"oneOf": [{"type": "null"}, {"$ref": "#/components/schemas/Error"}]}
        }
      },
      "Meta": {
        "type": "object",
        "required": ["contract_version", "schema_version", "snapshot_id", "as_of", "generated_at", "trace_id", "result_code", "partial", "missing_sections", "source_watermarks"],
        "properties": {
          "contract_version": {"type": "integer"},
          "schema_version": {"type": "integer"},
          "snapshot_id": {"type": "string"},
          "as_of": {"type": "string", "format": "date-time"},
          "generated_at": {"type": "string", "format": "date-time"},
          "trace_id": {"type": "string"},
          "result_code": {"type": "string"},
          "operation_id": {"type": "string"},
          "tenant_id": {"type": "string"},
          "partial": {"type": "boolean"},
          "missing_sections": {"type": "array", "items": {"type": "string"}},
          "source_watermarks": {"type": "object", "additionalProperties": {"type": "string"}},
          "projection_status": {"type": "string"}
        }
      },
      "Error": {
        "type": "object",
        "required": ["code", "message", "trace_id", "retryable"],
        "properties": {
          "code": {"type": "string"},
          "message": {"type": "string"},
          "trace_id": {"type": "string"},
          "retryable": {"type": "boolean"},
          "field_errors": {"type": "array", "items": {"type": "object", "required": ["field", "code", "message"], "properties": {"field": {"type": "string"}, "code": {"type": "string"}, "message": {"type": "string"}}}},
          "operation_id": {"type": "string"},
          "current_revision": {"type": "integer"},
          "projection_status": {"type": "string"},
          "status_url": {"type": "string"},
          "expected_consistency_window": {"type": "string"}
        }
      },
...
    }
  }
}